REFACTOR: to remove circular dependency
