+++ test/unit/services/discordService.test.js
@@ -5,14 +5,12 @@ const {
   addRoleToUser,
   getDiscordMembers,
   removeRoleFromUser,
-  archiveInactiveDiscordUsersInBulk,
 } = require("../../../services/discordService");
 const { fetchAllUsers } = require("../../../models/users");
 const Sinon = require("sinon");
 const userModel = firestore.collection("users");
 const userDataArray = require("../../fixtures/user/user")();
 const discordMembersArray = require("../../fixtures/discordResponse/discord-response");
-const cleanDb = require("../../utils/cleanDb");
 let fetchStub;
 describe("Discord services", function () {
   describe("setInDiscordFalseScript", function () {
@@ -112,68 +110,6 @@ describe("Discord services", function () {
       removeRoleFromUser("", "").catch((err) => {
         expect(err).to.be.an.instanceOf(Error);
         expect(err.message).to.equal("Fetch error");
-      });
-    });
-  });
-
-  describe("archive inactive discord users in bulk", function () {
-    const users = [];
-    beforeEach(async function () {
-      const addUsersPromises = [];
-      userDataArray.forEach((user) => {
-        const userData = {
-          ...user,
-          roles: {
-            ...user.roles,
-            in_discord: false,
-            archived: false,
-          },
-        };
-        addUsersPromises.push(userModel.add(userData));
-      });
-      await Promise.all(addUsersPromises);
-
-      users.length = 0;
-
-      const snapshot = await userModel
-        .where("roles.in_discord", "==", false)
-        .where("roles.archived", "==", false)
-        .get();
-
-      snapshot.forEach((user) => {
-        const id = user.id;
-        const userData = user.data();
-        users.push({ ...userData, id });
-      });
-    });
-
-    afterEach(async function () {
-      await cleanDb();
-      Sinon.restore();
-    });
-
-    it("Should return successful response", async function () {
-      const res = await archiveInactiveDiscordUsersInBulk(users);
-
-      expect(res).deep.equal({
-        message: "Successfully completed batch updates",
-        totalUsersArchived: 14,
-        totalOperationsFailed: 0,
-      });
-    });
-
-    it("should return failed response", async function () {
-      const batchStub = Sinon.stub(firestore, "batch");
-      batchStub.returns({
-        update: function () {},
-      });
-
-      const res = await archiveInactiveDiscordUsersInBulk(users);
-
-      expect(res).deep.equal({
-        message: "Firebase batch operation failed",
-        totalUsersArchived: 0,
-        totalOperationsFailed: 14,
       });
     });
   });
